+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+* Do not throw exceptions when a `frame.postMessage` is denied
 * Exceptions are no longer thrown when `publish`, `subscribe` or `unsubscribe` were invoked directly.
 
   ```javascript